+++ docs/Changelog.md
@@ -7,10 +7,16 @@ The format loosely follows *Keep a Changelog*, with a focus on user-visible DSP 
 ---
 
 ---
+
+## [v2.8.1.2] - 2026-01-27   
+
+### Fixed
+- Bug fix for modes selection, that was not saving ui state correctly
+
 ## [2.8.1.1] - 2026-01-27
 
 ### Ui-update
-- **Added modes selection (Basic & Advanced)**
+- Added modes selection (Basic & Advanced)
 
 ## [2.8.1] – 2026-01-25